refactor(recovery): extract format-state normalization helpers
